Feature/ci vs local env (#30)
* Working on tests (locally and on CI)

* Add quality tools and fix code

* Update config to make tests run on CI

* Add KERNAL_CLASS to phpunit config

* Fix style errors in tests

* Try ti fix tests on CI

* Still working on test issue on CI

* Adding env variable to tests execution

+++ src/Infrastructure/View/ViewHydrator/PaginationHydrator.php
@@ -6,10 +6,8 @@ use App\Infrastructure\View\ViewModel\PaginationViewModel;
 
 final class PaginationHydrator implements ViewHydratorInterface
 {
-
     public function __construct(private int $count, private int $page, private int $limit)
     {
-
     }
 
     public function hydrate(): array
@@ -17,9 +15,9 @@ final class PaginationHydrator implements ViewHydratorInterface
         $pagination = new PaginationViewModel(
             $this->count,
             $this->page,
-            ceil($this->count / $this->limit)
+            (int) ceil($this->count / $this->limit)
         );
 
-        return [ 'pagination' => $pagination];
+        return ['pagination' => $pagination];
     }
 }
